+++ extensions/sn_mod_support_apis/python/Send_Keys.py
@@ -78,7 +78,7 @@ if test_python_client:
     window_title = 'python.exe'
 
 
-def main():
+def main(args):
     '''
     Entry function for this server.
 
@@ -91,6 +91,11 @@ def main():
     Note: by experience, x4 side sometimes has several seconds of lag
     before processing keys sent.
     '''
+    # Enable test mode if requested.
+    if args['test']:
+        global test_python_client
+        test_python_client = True
+
     # Set up the pipe and connect to x4.
     pipe = Pipe_Server(pipe_name)